Rename fuzzer api package
