+++ evaluator/src/commonMain/kotlin/com/github/jcornaz/kwik/evaluator/FuzzerEvaluation.kt
@@ -1,7 +1,7 @@
 package com.github.jcornaz.kwik.evaluator
 
-import com.github.jcornaz.kwik.fuzzer.ExperimentalKwikFuzzer
-import com.github.jcornaz.kwik.fuzzer.Fuzzer
+import com.github.jcornaz.kwik.fuzzer.api.ExperimentalKwikFuzzer
+import com.github.jcornaz.kwik.fuzzer.api.Fuzzer
 import com.github.jcornaz.kwik.generator.api.randomSequence
