+++ evaluator/src/commonTest/kotlin/com/github/jcornaz/kwik/evaluator/ForAnyTest.kt
@@ -1,8 +1,8 @@
 package com.github.jcornaz.kwik.evaluator
 
-import com.github.jcornaz.kwik.fuzzer.ExperimentalKwikFuzzer
-import com.github.jcornaz.kwik.fuzzer.ensureAtLeastOne
-import com.github.jcornaz.kwik.fuzzer.toFuzzer
+import com.github.jcornaz.kwik.fuzzer.api.ExperimentalKwikFuzzer
+import com.github.jcornaz.kwik.fuzzer.api.ensureAtLeastOne
+import com.github.jcornaz.kwik.fuzzer.api.toFuzzer
 import com.github.jcornaz.kwik.generator.api.Generator
 import com.github.jcornaz.kwik.generator.api.randomSequence
 import com.github.jcornaz.kwik.generator.stdlib.ints
