+++ evaluator/src/jvmTest/kotlin/com/github/jcornaz/kwik/evaluator/DefaultIterationPropertyTest.kt
@@ -1,7 +1,7 @@
 package com.github.jcornaz.kwik.evaluator
 
-import com.github.jcornaz.kwik.fuzzer.ExperimentalKwikFuzzer
-import com.github.jcornaz.kwik.fuzzer.toFuzzer
+import com.github.jcornaz.kwik.fuzzer.api.ExperimentalKwikFuzzer
+import com.github.jcornaz.kwik.fuzzer.api.toFuzzer
 import com.github.jcornaz.kwik.generator.api.Generator
 import com.github.jcornaz.kwik.generator.stdlib.ints
 import kotlin.test.Test
