+++ fuzzer/api/src/commonMain/kotlin/com/github/jcornaz/kwik/fuzzer/api/ExperimentalKwikFuzzer.kt
@@ -1,4 +1,4 @@
-package com.github.jcornaz.kwik.fuzzer
+package com.github.jcornaz.kwik.fuzzer.api
 
 /**
  * The fuzzer system and related abstractions are extremely experimental and can be completely reworked between
